polovina modrho rozšíření, ale ještě trochu retardované

+++ app/app.py
@@ -38,13 +38,16 @@ class Jazyk(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False) 
 
-
+class Tags(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100), nullable=False)
+    barva = db.Column(db.String(100), nullable=False)
+    popis = db.Column(db.String(100), nullable=False)  
 
 class Kategorie(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100), nullable=False) 
-    barva = db.Column(db.String(100), nullable=False)
-    popis = db.Column(db.String(100), nullable=False)  
+    owned_id = db.Column(db.Integer, nullable=False)
+    type_id = db.Column(db.Integer, nullable=False)  
 
 class Filter():
     min_date = ""
@@ -70,6 +73,7 @@ def index():
 
 @app.route('/add/', methods=['POST', 'GET'])
 def addZaznam():  
+    tags = Tags.query.order_by(Tags.id).all()
     #pokud nekdo prida neco do databaze, tak se spusti tato cast, a pak se přeseměruje na view /zaznamy/
     record_name = str(request.form['name'])  
     record_time = int(request.form['time_spent'])
@@ -77,16 +81,26 @@ def addZaznam():
     record_jazyk = int(request.form['jazyk_id'])
     record_hodnoceni = int(request.form['hodnoceni'])
     record_date = str(request.form['date'])
-    
+
+
     new_record = Denik(name=record_name,jazyk_id=record_jazyk ,popis=record_popis,hodnoceni=record_hodnoceni,time_spent=record_time, date=record_date)
     db.session.add(new_record)
-    db.session.commit()
+    db.session.commit()    
+    for tag in tags:
+        try:
+            request.form[str(tag.id)]
+            new_cat = Kategorie(type_id=tag.id, owned_id=new_record.id)
+            db.session.add(new_cat)
+            db.session.commit()
+        except:
+            pass
     return redirect('/zaznamy/1')
 @app.route('/form/')
 def showForm():
     programmers = Programator.query.order_by(Programator.id).all()
     languages = Jazyk.query.order_by(Jazyk.id).all()
-    return render_template('addZaznam.html',languages=languages, programmers=programmers)
+    tags = Tags.query.order_by(Tags.id).all()
+    return render_template('addZaznam.html',languages=languages, programmers=programmers, tags=tags)
 
 @app.route('/language/')
 def showLanguageTable(): 
@@ -163,28 +177,37 @@ def delProgrammer(id):
 
 @app.route('/cat/')
 def showCatTable():
-    cats = Kategorie.query.order_by(Kategorie.id).all()
-    return render_template('kategorie.html',cats=cats)
+    tags = Kategorie.query.order_by(Kategorie.id).all()
+    return render_template('kategorie.html', tags=tags)
 
 @app.route('/cat/form/', methods=['POST', 'GET'])
 def showCatForm():
     if request.method == 'GET':
-        cats = Kategorie.query.order_by(Kategorie.id).all()
-        return render_template('addKategorie.html',cats=cats)
+        tags = Tags.query.order_by(Tags.id).all()
+        return render_template('addKategorie.html', tags=tags)
     else:
-        cat_name = request.form['name']
-        cat_type = request.form['type']
-        new_cat = Kategorie(name=cat_name, type=cat_type)
-        db.session.add(new_cat)
+        tag_name = request.form['name']
+        tag_barva = request.form['barva']
+        tag_popis = request.form['popis']
+
+        new_tag = Tags(name=tag_name, barva=tag_barva, popis=tag_popis)
+        db.session.add(new_tag)
         db.session.commit()
-        cats = Kategorie.query.order_by(Kategorie.id).all()
+        tags = Tags.query.order_by(Tags.id).all()
         request.method = "GET"
-        return render_template('addKategorie.html',cats=cats)
+        return render_template('addKategorie.html', tags=tags)
 
 @app.route('/cat/delete/<int:id>')
 def delCat(id):
-    cat_to_del = Kategorie.query.get_or_404(id)
-    db.session.delete(cat_to_del)
+    cats = Kategorie.query.order_by(Kategorie.id).all()
+    tag_to_del = Tags.query.get_or_404(id)
+    for cat in cats:
+        if (cat.type_id == tag_to_del.id):
+            cat_to_del = Kategorie.query.get_or_404(cat.id)
+            db.session.delete(cat_to_del)
+            db.session.commit()
+    
+    db.session.delete(tag_to_del)
     db.session.commit()
     return redirect('/cat/form/') 
 
@@ -194,6 +217,7 @@ def zaznamy(serazeni):
     languages = Jazyk.query.order_by(Jazyk.id).all()
     records = db.session.query(Denik).order_by(Denik.date).all()
     programmers = Programator.query.order_by(Programator.id).all()
+    tags = Tags.query.order_by(Tags.id).all()
     if(serazeni == 1):
         #datum od nejdříve
         records = db.session.query(Denik).order_by(Denik.date)
@@ -268,10 +292,12 @@ def zaznamy(serazeni):
     records = records.filter(Denik.time_spent >= Filter.time_from).filter(Denik.time_spent <= Filter.time_to)
     records = records.filter(Denik.hodnoceni <= Filter.hodnoceni_to).filter(Denik.hodnoceni >= Filter.hodnoceni_from)
     for language in languages:
-        if(not Filter.language_dict[language.id]):
+        if(not Filter.language_dict[int(language.id)]):
             records = records.filter(Denik.jazyk_id !=  language.id)
     
-    return render_template('zaznamy.html', records=records, languages=languages, cats=cats, programmers=programmers, filtered_languages=Filter.language_dict, min_date=Filter.date_from, max_date=Filter.date_to, min_time=Filter.time_from, max_time=Filter.time_to, max_hod=Filter.hodnoceni_to, min_hod=Filter.hodnoceni_from, sel_name=int(Filter.name), serazeni=serazeni)
+    return render_template('zaznamy.html', records=records, languages=languages, programmers=programmers, tags=tags, cats=cats, \
+    filtered_languages=Filter.language_dict, min_date=Filter.date_from, max_date=Filter.date_to, min_time=Filter.time_from, max_time=Filter. \
+    time_to, max_hod=Filter.hodnoceni_to, min_hod=Filter.hodnoceni_from, sel_name=int(Filter.name), serazeni=serazeni)
 
 @app.route('/zaznamy/set-serazeni/', methods=['POST', 'GET'])
 def setSerazeniZaznamy():
@@ -309,11 +335,19 @@ def OGzaznamy():
 def updateRecord(id):
     record_to_update = Denik.query.get_or_404(id)
     languages = Jazyk.query.order_by(Jazyk.id).all()
-    return render_template('update.html', record=record_to_update,languages=languages)
+    cats = Kategorie.query.order_by(Kategorie.id).all()
+    tags = Tags.query.order_by(Tags.id).all()
+    return render_template('update.html', record=record_to_update,languages=languages, tags=tags, cats=cats)
 
 @app.route('/delete/<int:id>')
 def deleteRecord(id):
     record_to_del = Denik.query.get_or_404(id)
+    cats = Kategorie.query.order_by(Kategorie.id).all()
+    for cat in cats:
+        if (cat.owned_id == record_to_del.id):
+            cat_to_del = Kategorie.query.get_or_404(cat.id)
+            db.session.delete(cat_to_del)
+            db.session.commit()
     db.session.delete(record_to_del)
     db.session.commit()
     return redirect('/zaznamy/1') 
@@ -327,6 +361,23 @@ def updateAddZaznam(id):
     record_to_update.jazyk_id = request.form['jazyk_id']
     record_to_update.hodnoceni = request.form['hodnoceni']
     record_to_update.date = request.form['date']
+    
+    cats = Kategorie.query.order_by(Kategorie.id).all()
+    for cat in cats:
+        if (cat.owned_id == record_to_update.id):
+            cat_to_del = Kategorie.query.get_or_404(cat.id)
+            db.session.delete(cat_to_del)
+            db.session.commit()
+
+    tags = Tags.query.order_by(Tags.id).all()
+    for tag in tags:
+        try:
+            request.form[str(tag.id)]
+            new_cat = Kategorie(type_id=tag.id, owned_id=record_to_update.id)
+            db.session.add(new_cat)
+            db.session.commit()
+        except:
+            pass
 
     db.session.commit()
     return redirect('/zaznamy/1')
